requirements.txt
gym~=0.21.0
torch~=1.10.1
opencv-python~=4.5.5.62
numpy~=1.22.0
Pillow~=9.0.0
torchvision~=0.11.2
matplotlib~=3.5.1
sklearn~=0.0
scikit-learn~=1.0.2
PyYAML~=6.0